Inplace weights loading
